requirements.txt
transformers==4.46.3
accelerate==1.2.1
bitsandbytes==0.44.1
safetensors==0.4.5
pyyaml==6.0.2
matplotlib==3.9.2
seaborn==0.13.2
pandas==2.2.3
# numpy is pulled by pandas/torch; pin only if you need reproducible resolution
pytest>=8.0.0
